Create collections index file and refactor imports
- Add src/collections/index.ts that exports all collections as an array
- Re-export each collection individually for named imports
- Update payload.config.ts to use centralized collections import
- Update test configuration to use centralized collections import
- Fix ESLint warnings (no-explicit-any and anonymous default export)

This centralizes collection management and makes it easier to add/remove
collections in the future by maintaining a single source of truth.
