+++ src/payload.config.ts
@@ -3,47 +3,46 @@ import { mongooseAdapter } from '@payloadcms/db-mongodb'
 import { payloadCloudPlugin } from '@payloadcms/payload-cloud'
 import { lexicalEditor } from '@payloadcms/richtext-lexical'
 import path from 'path'
-import { buildConfig } from 'payload'
+import { buildConfig, Config } from 'payload'
 import { fileURLToPath } from 'url'
 import sharp from 'sharp'
 
-import { Users } from './collections/Users'
-import { Media } from './collections/Media'
-import { Narrators } from './collections/Narrators'
-import { Meditations } from './collections/Meditations'
-import { Tags } from './collections/Tags'
-import { Music } from './collections/Music'
-import { Frames } from './collections/Frames'
+import { collections, Users } from './collections'
 
 const filename = fileURLToPath(import.meta.url)
 const dirname = path.dirname(filename)
 
-export default buildConfig({
-  admin: {
-    user: Users.slug,
-    importMap: {
-      baseDir: path.resolve(dirname),
+const payloadConfig = (overrides?: Config) => {
+  return buildConfig({
+    admin: {
+      user: Users.slug,
+      importMap: {
+        baseDir: path.resolve(dirname),
+      },
+      components: {
+        providers: [
+          {
+            path: './components/AdminProvider.tsx',
+          },
+        ],
+      },
     },
-    components: {
-      providers: [
-        {
-          path: './components/AdminProvider.tsx',
-        },
-      ],
+    collections,
+    editor: lexicalEditor(),
+    secret: process.env.PAYLOAD_SECRET || '',
+    typescript: {
+      outputFile: path.resolve(dirname, 'payload-types.ts'),
     },
-  },
-  collections: [Users, Media, Narrators, Meditations, Tags, Music, Frames],
-  editor: lexicalEditor(),
-  secret: process.env.PAYLOAD_SECRET || '',
-  typescript: {
-    outputFile: path.resolve(dirname, 'payload-types.ts'),
-  },
-  db: mongooseAdapter({
-    url: process.env.DATABASE_URI || '',
-  }),
-  sharp,
-  plugins: [
-    payloadCloudPlugin(),
-    // storage-adapter-placeholder
-  ],
-})
+    db: mongooseAdapter({
+      url: process.env.DATABASE_URI || '',
+    }),
+    sharp,
+    plugins: [
+      payloadCloudPlugin(),
+      // storage-adapter-placeholder
+    ],
+    ...overrides,
+  })
+}
+
+export default payloadConfig
